Added the basic stylizer.py file, it uses Magenta.
This was created for Bitcamp 2019, in about 8-9 hours at the end of the hackathon. I was trying to implement some other ideas with big data or fostering education, but I couldn't come up with an idea that I could implement in such a short time, so I went with this. I mostly learned tkinter and how to use Magenta.